+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to Pug will be documented in this file (beginning with v0.5 
 ### Added
 - Namespaces
 - Command support for namespaces (aka groups): enable, disable, remove, update
+- Rename projects
 
 ### Fixed
 - Always rebase, not just when changes are fetched